requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
groq>=0.4.0
elevenlabs>=1.0.0
